refactor: Balloon Game 로직을 custom hook으로 분리
